UTs on 64bit OS
git-svn-id: https://svn.code.sf.net/p/quex/code/trunk@4629 db9deaaf-66c7-40b0-811a-ad7aae109365

+++ quex/TESTS/drop-outs/TEST/test.c
@@ -44,7 +44,7 @@ main(int argc, char** argv)
 
     quex_EasyLexer_destruct(&qlex);
 
-    printf("[END] number of token = %i\n", number_of_tokens);
+    printf("[END] number of token = %i\n", (int)number_of_tokens);
     return 0;
 }
